fix bugs, push version from v0.11.1 to v0.11.2

+++ src/com/taobao/rigel/rap/project/dao/ProjectDao.java
@@ -78,7 +78,7 @@ public interface ProjectDao {
 	 * @param id
 	 * @return
 	 */
-	Action getAction(int id);
+	Action getAction(long id);
 
 	/**
 	 * save project
